requirements.txt
requests>=2.19.1
selenium>=3.14.0
numpy>=1.14.5
googletrans>=2.3.0
PyAutoGUI>=0.9.38
fake_useragent>=0.1.10
SpeechRecognition>=3.8.1
pydub>=0.22.1
ujson>=1.35
wiktionaryparser>=0.0.93
pandas>=0.23.4
textblob>=0.15.1
nltk>=3.2.1
matplotlib>=2.2.2
Pillow>=5.3.0
beautifulsoup4>=4.6.3
scikit_learn>=0.20.0
lxml